refactor(test): P5-1 extract animation API cluster from test_widget_bridge
Phase 5 P5-1 follow-up. Largest single-cluster extraction since the
initial canvas2d split — the entire bridge animation API surface
moved into test_widget_bridge_animation.cpp (25 TEST_CASEs, 120
assertions, 838 lines).

Covers the bridge ↔ MotionEngine plumbing:
  * setMotionToken from JS
  * Element.animate(...) + KeyframeEffect (Web Animations API)
  * animate() shorthand + motion provenance
  * pulp-motion-bench harness output

Duplicates the standard test helpers (trim_crlf, js_single_quoted,
wait_for_async_result) into the extracted TU to keep it self-contained
(established pattern from prior bundle-cuts on this file).

test_widget_bridge.cpp: 3,862 → 3,032 (-830 lines, under 5k and
within 10% of the 3,000-line P5-1 target).
